Add disability rating to profile
